+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `databaseSettings` parameter for `postSaveHook()` method
+- `databaseSettingsSource` getter to specify other settings source for client DB config
+
+### Fixed
+- `Client Create API` `postSaveHook()` client codes parameter
 
 ## [1.0.4] - 2020-05-04
 ### Fixed